browser/devtools/webide/test/test_fullscreenToolbox.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript;version=1.8" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8">

      function connectToLocal(win) {
        let deferred = promise.defer();
        win.AppManager.connection.once(
            win.Connection.Events.CONNECTED,
            () => deferred.resolve());
        win.document.querySelectorAll(".runtime-panel-item-other")[1].click();
        return deferred.promise;
      }

      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function* () {
          const { DebuggerServer } = require("devtools/server/main");
          Services.prefs.setBoolPref("devtools.webide.sidebars", false);
          win = yield openWebIDE();
          win.AppManager.update("runtime-list");

          yield connectToLocal(win);

          // Select main process
          yield win.Cmds.showProjectPanel();
          yield waitForUpdate(win, "runtime-targets");
          SimpleTest.executeSoon(() => {
            win.document.querySelectorAll("#project-panel-runtimeapps .panel-item")[0].click();
          });

          yield waitForUpdate(win, "project");

          ok(win.UI.toolboxPromise, "Toolbox promise exists");
          yield win.UI.toolboxPromise;

          ok(win.UI.toolboxIframe, "Toolbox iframe exists");

          let nbox = win.document.querySelector("#notificationbox");
          ok(nbox.hasAttribute("toolboxfullscreen"), "Toolbox is fullsreen");

          win.Cmds.showRuntimeDetails();

          ok(!nbox.hasAttribute("toolboxfullscreen"), "Toolbox is not fullscreen");

          yield win.Cmds.disconnectRuntime();

          yield closeWebIDE(win);

          DebuggerServer.destroy();

          SimpleTest.finish();
        });
      }


    </script>
  </body>
</html>
